fix: bullet-proof report generation — null safety + type casting sweep
Comprehensive fix across 15 files in the report pipeline:

**Null-safe snapshot access ($cur?-> instead of $cur->):**
- UptimeGatherer (4 lines), SecurityGatherer (1), BackupsGatherer (4),
  PerformanceGatherer (2), CloudflareGatherer (3), AnalyticsGatherer (2),
  SearchConsoleGatherer (3 snapshot + 5 cache collection access)

**Type casting for PostgreSQL string values:**
- BaseReportSectionGatherer: (float) cast in calculateTrend() arithmetic
  and formatNumber() — root cause of TypeError crashes
- AnalyticsGatherer: (int) cast for returning_users subtraction

**Nullable relationship safety:**
- ReportGeneratorService: reportConfig?->language, client?->name
- BackupsGatherer: config?->is_enabled, config?->frequency

**Blade template null guards:**
- analytics-1/2: array_column($a['devices/traffic_sources'] ?? [])
- search-console-2: collect($sc['overview']['daily_data'] ?? [])
- technical-stability: collect($db['categories'] ?? [])

**Other:**
- GenerateReport job: (int) filesize() cast
- WpUsersGatherer: (string) cast on email before explode()

+++ app/Jobs/GenerateReport.php
@@ -96,7 +96,7 @@ class GenerateReport implements ShouldBeUnique, ShouldQueue
                 ->update(['report_id' => $report->id]);
 
             $fullPath = Storage::disk('local')->path($filePath);
-            $fileSize = file_exists($fullPath) ? filesize($fullPath) : 0;
+            $fileSize = file_exists($fullPath) ? (int) filesize($fullPath) : 0;
             $fileName = basename($filePath);
 
             $report->update([
